third_party/blink/web_tests/images/animated-images-use-counter-001.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<body>
<script>
const kAnimatedImageUsedMoreThanOnce = 5605;

async_test(function(t) {
  internals.clearUseCounter(document, kAnimatedImageUsedMoreThanOnce);

  let gif1 = document.createElement("img");
  gif1.src = "resources/animated.gif";
  document.body.appendChild(gif1);
  let gif2 = document.createElement("img");
  gif2.src = "resources/animated.gif";
  document.body.appendChild(gif2);

  onload = t.step_func(() => {
    requestAnimationFrame(t.step_func(() => {
      requestAnimationFrame(t.step_func_done(() => {
        assert_true(internals.isUseCounted(document, kAnimatedImageUsedMoreThanOnce));
      }));
    }));
  });
}, 'Check that kAnimatedImageUsedMoreThanOnce is counted with a duplicated animated image');
</script>
</body>
